Addressing feedback

+++ src/L6/datasorter_test.py
@@ -27,8 +27,13 @@ class DataSorterTest(unittest.TestCase):
 
     def test_set_input_data_with_valid_paths(self):
         self.assertTrue(self.under_test.set_input_data("test2.csv"))
+        self.assertIsNotNone(self.under_test.data)
+        self.assertIsInstance(self.under_test.data, list)
+        self.assertEqual(3, len(self.under_test.data))
+        self.assertEqual('3', self.under_test.data[2])
 
     def tearDown(self):
-        os.remove("test1.txt")
-        os.remove("test2.csv")
-        pass
+        if os.path.exists("test1.txt"):
+            os.remove("test1.txt")
+        if os.path.exists("test2.csv"):
+            os.remove("test2.csv")
